% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/election.R
\docType{class}
\name{election-class}
\alias{election-class}
\alias{election}
\title{Election in Sweden}
\arguments{
\item{city_name}{object of class 'election': search the result of election according to city.}
}
\value{
Returns an object of class 'election'.
}
\description{
Votes for the different parties in the election in Sweden.
Also, first-time voters who voted in the election to the Riksdag.
Voter turnout in the parliamentary election.
}
\section{Fields}{

\describe{
\item{\code{city_name}}{character.}

\item{\code{kpi}}{character.}

\item{\code{id}}{character.}

\item{\code{year}}{character.}

\item{\code{participation_rate_data}}{data.frame.}

\item{\code{political_party_data}}{data.frame.}

\item{\code{first_vote_data}}{data.frame.}
}}


\examples{
el_object <- election("Stockholm")
el_object$participation_rate()
el_object$political_party()
el_object$first_vote()

}
