% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val2014.R
\name{val2014}
\alias{val2014}
\title{Election distribution plot.}
\usage{
val2014(val_data, lan_val, kommun_val)
}
\arguments{
\item{val_data}{A number to choose election type}

\item{lan_val}{A number to choose county}

\item{kommun_val}{A number to choose city}
}
\value{
A plot
}
\description{
Election distribution plot.
}
